+++ CHANGES.md
@@ -1,4 +1,6 @@
 
+  - fixed division by zero in rotations of 0-length vectors
+    (reported by Nikolaus Huber)
   - fixed integer overflows in bound tests in `fill`, `blit`, `sub`
     (fix by Nikolaus Huber)