+++ CHANGES.md
@@ -1,4 +1,5 @@
 
+  - new in-place bitwise operations (contribution by Mathieu Barbin)
   - fixed division by zero in rotations of 0-length vectors
     (reported by Nikolaus Huber)
   - fixed integer overflows in bound tests in `fill`, `blit`, `sub`